Initial commit of code.
This is the initial commit of code from the files stored in the ACM
Digital Library (oto-test-06102005.tgz and graphgen-06102005.tgz).
These contain, as far as I know, the very latest version.  I also had a
subversion repository which would have contained history.  But, this is
currently corrupted and I need to figure out what I can do there (if
anything).