Adding missing code for batch MNR algorithm
This algorithm is code for the paper entitled: "A Batch Algorithm for
Maintaining a Topological Order", published at ACSC in 2010.  For some
reason, the code was never added to this repository.  Unfortunately, due
to bit rot, its unlikely it compiles.  However, it is at least here now
for the record.